+++ tests/testthat/test_analysis_example_ishak2007.r
@@ -26,7 +26,7 @@ V <- V[!is.miss,!is.miss] ### remove missing measurement occasions from V
 
 test_that("results are correct for diag(V) and struct='DIAG'.", {
 
-   res <- rma.mv(yi, diag(V), mods = ~ factor(time) - 1, random = ~ factor(time) | study,
+   res <- rma.mv(yi, diag(V), mods = ~ 0 + factor(time), random = ~ factor(time) | study,
                  struct = "DIAG", data = dat.long, sparse=.sparse)
 
    ### Table 1, column "Time-specific (Independence)"
@@ -37,7 +37,7 @@ test_that("results are correct for diag(V) and struct='DIAG'.", {
 
 test_that("results are correct for diag(V) and random study effects.", {
 
-   res <- rma.mv(yi, diag(V), mods = ~ factor(time) - 1, random = ~ 1 | study, data = dat.long, sparse=.sparse)
+   res <- rma.mv(yi, diag(V), mods = ~ 0 + factor(time), random = ~ 1 | study, data = dat.long, sparse=.sparse)
 
    ### Table 1, column "Random study effects"
    expect_equivalent(coef(res), c(-26.2127, -27.1916, -28.5464, -25.6339), tolerance=.tol[["coef"]])
@@ -47,7 +47,7 @@ test_that("results are correct for diag(V) and random study effects.", {
 
 test_that("results are correct for diag(V) and struct='ID'.", {
 
-   res <- rma.mv(yi, diag(V), mods = ~ factor(time) - 1, random = ~ factor(time) | study,
+   res <- rma.mv(yi, diag(V), mods = ~ 0 + factor(time), random = ~ factor(time) | study,
                  struct = "ID", data = dat.long, sparse=.sparse)
 
    ### not in paper
@@ -58,7 +58,7 @@ test_that("results are correct for diag(V) and struct='ID'.", {
 
 test_that("results are correct for diag(V) and struct='HAR'.", {
 
-   res <- rma.mv(yi, diag(V), mods = ~ factor(time) - 1, random = ~ time | study,
+   res <- rma.mv(yi, diag(V), mods = ~ 0 + factor(time), random = ~ time | study,
                  struct = "HAR", data = dat.long, sparse=.sparse)
 
    ### Table 1, column "Correlated random time effects"
@@ -70,7 +70,7 @@ test_that("results are correct for diag(V) and struct='HAR'.", {
 
 test_that("results are correct for struct='HAR'.", {
 
-   res <- rma.mv(yi, V, mods = ~ factor(time) - 1, random = ~ time | study,
+   res <- rma.mv(yi, V, mods = ~ 0 + factor(time), random = ~ time | study,
                  struct = "HAR", data = dat.long, sparse=.sparse)
 
    ### Table 1, column "Multivariate model"
@@ -82,7 +82,7 @@ test_that("results are correct for struct='HAR'.", {
 
 test_that("results are correct for struct='AR'.", {
 
-   res <- rma.mv(yi, V, mods = ~ factor(time) - 1, random = ~ time | study,
+   res <- rma.mv(yi, V, mods = ~ 0 + factor(time), random = ~ time | study,
                  struct = "AR", data = dat.long, sparse=.sparse)
 
    ### not in paper
@@ -94,7 +94,7 @@ test_that("results are correct for struct='AR'.", {
 
 test_that("results are correct for struct='HCS'.", {
 
-   res <- rma.mv(yi, V, mods = ~ factor(time) - 1, random = ~ factor(time) | study,
+   res <- rma.mv(yi, V, mods = ~ 0 + factor(time), random = ~ factor(time) | study,
                  struct = "HCS", data = dat.long, sparse=.sparse)
 
    ### not in paper
@@ -105,7 +105,7 @@ test_that("results are correct for struct='HCS'.", {
 
 test_that("results are correct for struct='CAR'.", {
 
-   res <- rma.mv(yi, V, mods = ~ factor(time) - 1, random = ~ time | study,
+   res <- rma.mv(yi, V, mods = ~ 0 + factor(time), random = ~ time | study,
                  struct = "CAR", data = dat.long, sparse=.sparse)
 
    ### not in paper
@@ -122,7 +122,7 @@ test_that("results are correct for struct='CAR' with unequally spaced time point
    dat.long$time[dat.long$time == 2] <-  6/3
    dat.long$time[dat.long$time == 1] <-  3/3
 
-   res <- rma.mv(yi, V, mods = ~ factor(time) - 1, random = ~ time | study,
+   res <- rma.mv(yi, V, mods = ~ 0 + factor(time), random = ~ time | study,
                  struct = "CAR", data = dat.long, sparse=.sparse)
 
    ### not in paper