+++ tests/testthat/test_misc_rma_mv.r
@@ -126,12 +126,12 @@ test_that("rma.mv() correctly handles 'beta' argument", {
 
    dat <- dat.berkey1998
    V <- vcalc(vi=1, cluster=author, rvars=c(v1i, v2i), data=dat)
-   res.unc <- rma.mv(yi, V, mods = ~ outcome - 1, random = ~ outcome | trial, struct="UN", data=dat, method="ML")
-   res.01 <- rma.mv(yi, V, mods = ~ outcome - 1, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(0,0))
-   res.02 <- rma.mv(yi, V, mods = ~ outcome - 1, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(NA,0))
-   res.03 <- rma.mv(yi, V, mods = ~ outcome - 1, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(0,NA))
+   res.un <- rma.mv(yi, V, mods = ~ 0 + outcome, random = ~ outcome | trial, struct="UN", data=dat, method="ML")
+   res.01 <- rma.mv(yi, V, mods = ~ 0 + outcome, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(0,0))
+   res.02 <- rma.mv(yi, V, mods = ~ 0 + outcome, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(NA,0))
+   res.03 <- rma.mv(yi, V, mods = ~ 0 + outcome, random = ~ outcome | trial, struct="UN", data=dat, method="ML", beta=c(0,NA))
 
-   fstats <- fitstats(res.01, res.02, res.03, res.unc)
+   fstats <- fitstats(res.01, res.02, res.03, res.un)
    expect_equivalent(unlist(fstats[1,]), c(-2.464111, -0.691524, 1.010033, 5.840657), tolerance=.tol[["fit"]])
 
    dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)