+++ NEWS.md
@@ -1,12 +1,16 @@
-# metafor 4.7-14 (2024-06-07)
+# metafor 4.7-15 (2024-06-10)
+
+- the `predict.rma()` and `predict.rma.ls()` functions now also accept a matrix as input that includes a column for the intercept term (in which case the `intercept` argument is ignored)
+
+- added `pairwise()` function to construct a matrix of pairwise contrasts
 
 - `rma.mv()` now counts the number of levels of a random effect more appropriately; this may trigger more often the check whether the number of levels is equal to 1, in which case the corresponding variance component is automatically fixed to 0; this check can be omitted with `control=list(check.k.gtr.1=FALSE)`
 
 - made optimizers `Rcgmin` and `Rvmmin` available again via the `optimx` package
 
-- argument `shade` in `funnel()` now automatically uses a color gradient for the regions when multiple `level` values are specified
+- when unspecified, argument `shade` in `funnel()` now automatically uses a color gradient for the regions when multiple `level` values are specified
 
-- added extractor for function `se()` for extracting standard errors from model objects
+- added extractor function `se()` for extracting standard errors from model objects
 
 - added `lim`, `ci`, `pi`, `legend`, and `flip` arguments to `labbe()`